requirements.txt
flask
requests
websocket_client
redis
peewee
flask_socketio
flask_cors